Converted maxBalance to decimal field

+++ MT/forms.py
@@ -9,5 +9,5 @@ class userTimezoneForm(forms.Form):
     timezone = forms.ChoiceField(choices=[(x, x) for x in pytz.common_timezones])
 
 class configForm (forms.Form):
-    configMaxBet = forms.IntegerField(widget=forms.NumberInput)
+    configMaxBet = forms.DecimalField(widget=forms.NumberInput,decimal_places=2)
     configProcessEnabled = forms.BooleanField(widget=forms.CheckboxInput,required=False)
